+++ PubMatic/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 4.7.0.0
+* Certified with PubMatic SDK 4.7.0.
+* Updated to use PubMatic's new `loadAd` API that includes MAX as the bidding host.
+* Updated ad display failed error code.
+
 ## 4.6.0.0
 * Certified with PubMatic SDK 4.6.0.